refactor logical plan generator
